+++ lib/cheftacular/helper.rb
@@ -130,7 +130,7 @@ class Cheftacular
       user     = @config['cheftacular']['deploy_user']
       password = @config['server_passwords'][@options['address']]
       nodename = @options['node_name']
-      chef_ver = @config['cheftacular']['chef_client_version']
+      chef_ver = @config['cheftacular']['chef_version'].to_i >= 12 ? '11.16.4' : '12.4.0'
 
       "knife bootstrap #{ address } -x #{ user } -P #{ password } -N #{ nodename } --sudo --use-sudo-password --bootstrap-version #{ chef_ver }"
     end
@@ -166,11 +166,11 @@ class Cheftacular
     end
 
     def set_cloud_options
-      @options['preferred_cloud']        = @options['preferred_cloud'].nil? ?        @config['cheftacular']['preferred_cloud'] :        @options['preferred_cloud']
-      @options['preferred_cloud_image']  = @options['preferred_cloud_image'].nil? ?  @config['cheftacular']['preferred_cloud_image'] :  @options['preferred_cloud_image']
-      @options['preferred_cloud_region'] = @options['preferred_cloud_region'].nil? ? @config['cheftacular']['preferred_cloud_region'] : @options['preferred_cloud_region']
-      @options['virtualization_mode']    = @options['virtualization_mode'].nil? ?    @config['cheftacular']['virtualization_mode'] :    @options['virtualization_mode']
-      @options['route_dns_changes_via']  = @options['route_dns_changes_via'].nil? ?  @config['cheftacular']['route_dns_changes_via'] :  @options['route_dns_changes_via']
+      @options['preferred_cloud']        = @options['preferred_cloud'].nil? ?        @config['cheftacular']['preferred_cloud'].downcase : @options['preferred_cloud'].downcase
+      @options['preferred_cloud_image']  = @options['preferred_cloud_image'].nil? ?  @config['cheftacular']['preferred_cloud_image'] :    @options['preferred_cloud_image']
+      @options['preferred_cloud_region'] = @options['preferred_cloud_region'].nil? ? @config['cheftacular']['preferred_cloud_region'] :   @options['preferred_cloud_region']
+      @options['virtualization_mode']    = @options['virtualization_mode'].nil? ?    @config['cheftacular']['virtualization_mode'] :      @options['virtualization_mode']
+      @options['route_dns_changes_via']  = @options['route_dns_changes_via'].nil? ?  @config['cheftacular']['route_dns_changes_via'] :    @options['route_dns_changes_via']
     end
 
     def does_cheftacular_config_have? key_array