+++ lib/cheftacular/stateless_actions/remove_client.rb
@@ -3,10 +3,14 @@ class Cheftacular
   class StatelessActionDocumentation
     def remove_client
       @config['documentation']['stateless_action'] <<  [
-        "`cft remove_client -n NODE_NAME` removes a client (and its node data) from the chef-server. " +
+        "`cft remove_client NODE_NAME [destroy]` removes a client (and its node data) from the chef-server. " +
         "It also removes its dns records from the cloud service (if possible). " +
         "This should not be done lightly as you will have to wipe the server and trigger another chef-client " +
-        "run to get it to register again. Alternatively, you can run `cft reinitialize IP_ADDRESS NODE_NAME as well."
+        "run to get it to register again. Alternatively, you can run `cft reinitialize IP_ADDRESS NODE_NAME as well.",
+
+        [
+          "    1. `destroy` deletes the server as well as removing it from the chef environment."
+        ]
       ]
     end
   end