+++ lib/cloud_interactor/server/create.rb
@@ -27,8 +27,9 @@ class CloudInteractor
       }
 
       if @options['preferred_cloud'] == 'digitalocean'
-        final_create_args[:region_id]   = @main_obj['specific_regions'].first['id']
-        final_create_args[:ssh_key_ids] = @main_obj['specific_ssh_keys'].first['id']
+        final_create_args[:region_id]          = @main_obj['specific_regions'].first['id']
+        final_create_args[:ssh_key_ids]        = @main_obj['specific_ssh_keys'].first['id']
+        final_create_args[:private_networking] = true
       end
 
       @main_obj["#{ IDENTITY }_create_request"] = JSON.parse(@classes['auth'].auth_service(RESOURCE).instance_eval(IDENTITY).create(final_create_args).to_json)